requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
APScheduler>=3.9.1
asyncpg>=0.27.0